Move Most GDAL-Dependents into geotrellis-gdal
On Ubuntu 14.04, I had to build GDAL and the Java bindings by
downloading http://download.osgeo.org/gdal/2.0.2/gdal-2.0.2.tar.xz,
untarring, and typing

```shell
./configure --prefix=$HOME/local/gdal-2.0.2 --with-java && make
cd swig/java
make
```

then building/running tests by typing

```shell
LD_LIBRARY_PATH=$HOME/local/src/gdal-2.0.2/swig/java:$LD_LIBRARY_PATH ./buildall.sh
```
