+++ build.sbt
@@ -136,7 +136,7 @@ lazy val index = Project("index", file("index")).
   settings(commonSettings: _*)
 
 lazy val gdal: Project = Project("gdal", file("gdal")).
-  dependsOn(raster, geotools % "test").
+  dependsOn(proj4, raster, spark, geotools % "test").
   settings(commonSettings: _*)
 
 lazy val geotools = Project("geotools", file("geotools")).
