+++ gdal/src/main/scala/geotrellis/gdal/io/formats/NetCdfInputFormat.scala
@@ -1,4 +1,4 @@
-package geotrellis.spark.io.hadoop.formats
+package geotrellis.gdal.io.formats
 
 import org.joda.time.{DateTimeZone, DateTime}
