+++ gdal/src/test/scala/geotrellis/gdal/io/formats/NetCdfInputFormatSpec.scala
@@ -1,4 +1,4 @@
-package geotrellis.spark.io.hadoop.formats
+package geotrellis.gdal.io.formats
 
 import org.scalatest.{FunSpec, Matchers}
